+++ lib/f_service/rspec/support.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+require_relative 'support/helpers'
+require_relative 'support/matchers'